+++ common-system-test/src/main/java/com/teamscale/test/commons/SystemTestUtils.java
@@ -1,12 +1,16 @@
 package com.teamscale.test.commons;
 
 import com.teamscale.report.testwise.model.TestInfo;
+import okhttp3.MediaType;
+import okhttp3.RequestBody;
 import org.apache.commons.lang3.SystemUtils;
 import org.conqat.lib.commons.io.ProcessUtils;
 import org.jetbrains.annotations.NotNull;
 import retrofit2.Call;
 import retrofit2.Retrofit;
+import retrofit2.http.Body;
 import retrofit2.http.POST;
+import retrofit2.http.PUT;
 
 import java.io.File;
 import java.io.IOException;
@@ -150,12 +154,27 @@ public class SystemTestUtils {
 		/** Dumps coverage */
 		@POST("/dump")
 		Call<Void> dump();
+
+		/** Changes the current partition. */
+		@PUT("/partition")
+		Call<Void> changePartition(@Body RequestBody newPartition);
+
+		/** Changes the current partition. Convenience method to pass a plain string. */
+		default Call<Void> changePartition(String newPartition) {
+			return changePartition(RequestBody.create(MediaType.parse("text/plain"), newPartition));
+		}
 	}
 
 	/** Instructs the agent via HTTP to dump the currently collected coverage. */
 	public static void dumpCoverage(int agentPort) throws IOException {
 		new Retrofit.Builder().baseUrl("http://localhost:" + agentPort).build()
 				.create(AgentService.class).dump().execute();
+	}
+
+	/** Instructs the agent via HTTP to change the partition. */
+	public static void changePartition(int agentPort, String newPartition) throws IOException {
+		new Retrofit.Builder().baseUrl("http://localhost:" + agentPort).build()
+				.create(AgentService.class).changePartition(newPartition).execute();
 	}
 
 }