+++ system-tests/gradle-multi-module/gradle-project/module1/src/test/java/org/example/Test1.java
@@ -1,0 +1,11 @@
+package org.example;
+
+import org.junit.jupiter.api.Test;
+
+public class Test1 {
+
+    @Test
+    void test1() {
+        Main1.main(new String[]{});
+    }
+}